requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
azure-ai-projects==1.0.0b5
azure-core==1.32.0
azure-identity==1.19.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cloudevents==1.11.0
colorama==0.4.6
cryptography==44.0.0
defusedxml==0.7.1
Deprecated==1.2.15
deprecation==2.1.0
distro==1.9.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
isodate==0.7.2
Jinja2==3.1.5
jiter==0.8.2
jsonschema==4.23.0
jsonschema-path==0.3.3
jsonschema-specifications==2024.10.1
lazy-object-proxy==1.10.0
MarkupSafe==3.0.2
more-itertools==10.6.0
msal==1.31.1
msal-extensions==1.2.0
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.2.2
openai==1.60.0
openapi-core==0.19.4
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.1
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
packaging==24.2
parse==1.20.2
pathable==0.4.4
portalocker==2.10.1
prance==23.6.21.0
propcache==0.2.1
pybars4==0.9.13
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyJWT==2.10.1
PyMeta3==0.5.1
python-dotenv==1.0.1
pywin32==308
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rpds-py==0.22.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
semantic-kernel==1.19.0
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0